requirements.txt
apsw
Flask
pycryptodome
pyotp
eel
